Fix: Clear order state when adding an account
We recently stopped reseting order state when getting an error in the checkout process so that you could go back and fix any issues. This is causing real orders to be created when adding an account if you previously tried to give but were unsuccessful. This should fix that by clearing order state when you go to add an account.